pyproject.toml
[tool.poetry]
name = "document-search"
version = "0.1.0"
description = ""
authors = ["Nikhil Maheshwari"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.45.1"
pillow = "^11.2.1"
pytesseract = "^0.3.13"
pypdf2 = "^3.0.1"
pymupdf = "^1.26.1"
litellm = "^1.72.4"
python-dotenv = "^1.1.0"
qdrant-client = "^1.14.2"
openpyxl = "^3.1.5"
langchain-text-splitters = "^0.3.8"
langchain-core = ">=0.3.0"
colorlog = "^6.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
